backend/requirements.txt
# Web框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.25
aiosqlite==0.19.0

# 向量数据库
chromadb==0.4.22

# AI服务
openai==1.12.0
httpx==0.26.0

# 数据处理
pyyaml==6.0.1
python-dotenv==1.0.1
tiktoken==0.5.2

# 工具库
pydantic==2.6.0
pydantic-settings==2.1.0
uuid6==2024.1.12

# 日志
loguru==0.7.2

